+++ test/resources/unused_code_analyzer/conditional_prefixed_file.dart
@@ -1,0 +1,3 @@
+int calculate() => 1;
+
+bool helloWorld() => false; // LINT