+++ test/resources/unused_code_analyzer/unconditional_prefixed_file.dart
@@ -1,0 +1,1 @@
+int calculate() => 2;